audit fix: corrected the typecasting order that prevents over/under flow

+++ src/proposal/base/ModuleManager.sol
@@ -64,8 +64,7 @@ abstract contract ModuleManager is
     }
 
     modifier moduleLimitNotExceeded() {
-        // @audit susceptible to silent overflow
-        if (uint8(_modules.length) >= MAX_MODULE_AMOUNT) {
+        if (_modules.length >= uint256(MAX_MODULE_AMOUNT)) {
             revert Proposal__ModuleManager__ModuleAmountOverLimits();
         }
         _;
